Add response check to Past and Upcoming Conference Tests. Modify Upcoming Conferences Tests
