+++ qa/api/src/test/resources/config.properties
@@ -1,4 +1,5 @@
-baseHost=http://localhost:8025
+#baseHost=http://localhost:8025
+baseHost=http://ecsd00100875.epam.com:8025
 
 
 # User credentials
